requirements.txt
fake-useragent==0.1.11
Flask==2.0.2
Flask-WTF==1.0.0
gunicorn==20.1.0
kaggle==1.5.12
matplotlib==3.1.3
numpy==1.21.2
optuna==2.10.0
Pillow==7.1.2
python-dotenv==0.19.2
seaborn==0.11.2
selenium==3.141.0
scikit-learn==1.0
requests==2.24.0
retry==0.9.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
ttach==0.0.3
varname==0.8.1
WTForms==3.0.1
